Add rich weapon system

+++ Assets/Scripts/ECS/Components/Team.cs
@@ -6,5 +6,5 @@ using Unity.Entities;
 [GenerateAuthoringComponent]
 public struct Team : IComponentData
 {
-    public int Id;
+    public byte Id;
 }
